requirements.txt
pymilvus>=2.4.0
openai>=1.40.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
streamlit>=1.38.0
tiktoken>=0.7.0
PyPDF2>=3.0.1
numpy>=1.24.0
pandas>=2.0.0 